backend/lambda/parse_pages/requirements.txt
requests
beautifulsoup4
usaddress-scourgify